implemented library selection

+++ tcamViewer/src/main/java/com/darcangel/tcamViewer/container/LibraryItemDetails.java
@@ -7,15 +7,15 @@ import androidx.annotation.Nullable;
 import androidx.recyclerview.selection.ItemDetailsLookup;
 
 public class LibraryItemDetails extends ItemDetailsLookup.ItemDetails<Long> {
-    private int position;
+    private Long position;
 
-    public LibraryItemDetails(final int position) {
+    public LibraryItemDetails(final Long position) {
         this.position = position;
     }
 
     @Override
     public int getPosition() {
-        return position;
+        return position.intValue();
     }
 
     @Nullable
